% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello1.R
\name{qplot_001}
\alias{qplot_001}
\title{This is example 2.}
\usage{
qplot_001(a, b)
}
\arguments{
\item{a}{must be ib function.}

\item{b}{if a must b.}
}
\value{
mtcars
}
\description{
\code{add} an example 2.
}
\details{
Which is an example.
}
